# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Basketball attendance and player evaluation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface UserMe {
  id: number;
  username: string;
  fullName: string;
  role: string;
  isActive: boolean;
}

export interface Coach {
  id: number;
  username: string;
  fullName: string;
  role: string;
  isActive: boolean;
}

export interface CoachInput {
  username: string;
  fullName: string;
  password: string;
}

export interface CoachUpdate {
  fullName?: string;
  password?: string;
}

export interface Team {
  id: number;
  name: string;
  /** @nullable */
  ageCategory?: string | null;
  /** @nullable */
  coachId?: number | null;
  /** @nullable */
  coachName?: string | null;
}

export interface TeamInput {
  name: string;
  ageCategory?: string;
  /** @nullable */
  coachId?: number | null;
}

export interface TeamUpdate {
  name?: string;
  ageCategory?: string;
  /** @nullable */
  coachId?: number | null;
}

export interface Player {
  id: number;
  name: string;
  teamId: number;
  /** @nullable */
  position?: string | null;
  /** @nullable */
  jerseyNumber?: number | null;
  isActive: boolean;
}

export interface PlayerInput {
  name: string;
  position?: string;
  /** @nullable */
  jerseyNumber?: number | null;
}

export interface PlayerUpdate {
  name?: string;
  position?: string;
  /** @nullable */
  jerseyNumber?: number | null;
}

export interface TrainingSession {
  id: number;
  teamId: number;
  date: string;
  /** @nullable */
  notes?: string | null;
  attendanceCount?: number;
}

export interface SessionInput {
  date: string;
  notes?: string;
}

export interface PlayerEvaluation {
  criterionId: number;
  criterionName: string;
  score: number;
}

export interface PlayerAttendance {
  playerId: number;
  playerName: string;
  /** @nullable */
  jerseyNumber?: number | null;
  status: string;
  /** @nullable */
  note?: string | null;
  evaluations?: PlayerEvaluation[];
}

export interface Criterion {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  isActive: boolean;
}

export interface SessionDetail {
  id: number;
  teamId: number;
  date: string;
  /** @nullable */
  notes?: string | null;
  players: PlayerAttendance[];
  criteria: Criterion[];
}

export interface EvaluationScore {
  criterionId: number;
  score: number;
}

export interface AttendanceEntry {
  playerId: number;
  status: string;
  note?: string;
  scores?: EvaluationScore[];
}

export interface AttendanceInput {
  entries: AttendanceEntry[];
}

export interface CriterionInput {
  name: string;
  description?: string;
}

export interface CriterionAvg {
  criterionId: number;
  criterionName: string;
  /** @nullable */
  avg?: number | null;
}

export interface PlayerReport {
  playerId: number;
  playerName: string;
  /** @nullable */
  jerseyNumber?: number | null;
  totalSessions: number;
  present: number;
  absent: number;
  late: number;
  excused: number;
  attendancePct: number;
  /** @nullable */
  overallAvg?: number | null;
  criteriaAvgs?: CriterionAvg[];
}

export interface TeamReport {
  teamId: number;
  teamName: string;
  totalSessions: number;
  criteria: Criterion[];
  players: PlayerReport[];
}

export interface ReportFilter {
  dateFrom?: string;
  dateTo?: string;
}

export interface TeamStat {
  teamId: number;
  teamName: string;
  /** @nullable */
  ageCategory?: string | null;
  /** @nullable */
  coachName?: string | null;
  sessionsCount: number;
  playersCount: number;
}

export interface DashboardStats {
  totalTeams: number;
  totalCoaches: number;
  totalPlayers: number;
  totalSessions: number;
  teams: TeamStat[];
}