# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Basketball attendance and player evaluation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserMe"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserMe"
        "401":
          description: Not authenticated
  /coaches:
    get:
      operationId: listCoaches
      tags: [coaches]
      summary: List all coaches
      responses:
        "200":
          description: List of coaches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coach"
    post:
      operationId: createCoach
      tags: [coaches]
      summary: Create a coach (head_coach only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoachInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coach"
  /coaches/{id}:
    patch:
      operationId: updateCoach
      tags: [coaches]
      summary: Update a coach
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoachUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coach"
    delete:
      operationId: deleteCoach
      tags: [coaches]
      summary: Delete a coach
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /coaches/{id}/toggle:
    patch:
      operationId: toggleCoach
      tags: [coaches]
      summary: Toggle coach active status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coach"
...
  /teams/{id}/players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List players in a team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
    post:
      operationId: createPlayer
      tags: [players]
      summary: Add a player to a team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/{id}:
    patch:
      operationId: updatePlayer
      tags: [players]
      summary: Update a player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    delete:
      operationId: deletePlayer
      tags: [players]
      summary: Delete a player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /players/{id}/toggle:
    patch:
      operationId: togglePlayer
      tags: [players]
      summary: Toggle player active status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
...
  /criteria:
    get:
      operationId: listCriteria
      tags: [criteria]
      summary: List evaluation criteria
      responses:
        "200":
          description: List of criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Criterion"
    post:
      operationId: createCriterion
      tags: [criteria]
      summary: Create an evaluation criterion (head_coach only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriterionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Criterion"
  /criteria/{id}/toggle:
    patch:
      operationId: toggleCriterion
      tags: [criteria]
      summary: Toggle criterion active status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Criterion"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    UserMe:
      type: object
      required: [id, username, fullName, role, isActive]
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        role:
          type: string
        isActive:
          type: boolean
    Coach:
      type: object
      required: [id, username, fullName, role, isActive]
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        role:
          type: string
        isActive:
          type: boolean
    CoachInput:
      type: object
      required: [username, fullName, password]
      properties:
        username:
          type: string
        fullName:
          type: string
        password:
          type: string
    CoachUpdate:
      type: object
      properties:
        fullName:
          type: string
        password:
          type: string
...
    Player:
      type: object
      required: [id, name, teamId, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        teamId:
          type: integer
        position:
          type: ["string", "null"]
        jerseyNumber:
          type: ["integer", "null"]
        isActive:
          type: boolean
    PlayerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        position:
          type: string
        jerseyNumber:
          type: ["integer", "null"]
    PlayerUpdate:
      type: object
      properties:
        name:
          type: string
        position:
          type: string
        jerseyNumber:
          type: ["integer", "null"]
...
    Criterion:
      type: object
      required: [id, name, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
    CriterionInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
...