# lib/api-zod/src/generated/types/coachInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Basketball attendance and player evaluation API
 * OpenAPI spec version: 0.1.0
 */

export interface CoachInput {
  username: string;
  fullName: string;
  password: string;
}