# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Basketball attendance and player evaluation API
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  id: number;
  name: string;
  teamId: number;
  /** @nullable */
  position?: string | null;
  /** @nullable */
  jerseyNumber?: number | null;
  isActive: boolean;
}